Initial setup + clickable tiles
I didn't want to use MeshColliders, but they're a lot faster than
rotating three box colliders to fill the hex. The mesh is simple enough
that I'm fine with it.
There are a lot of magic numbers in InitializePieces(). I can't think of
a better place to hide them, since they really are just arbitrary
numbers for where the pieces should start. If someone ever reads this
and knows a better way to do this, I'm all ears.